Split into files of RegisterAllocator
