+++ src/error.rs
@@ -4,8 +4,8 @@ use crate::analyze::expr::ConvExprKind;
 use crate::analyze::expr::FuncCallTargetKind;
 use crate::analyze::types::Type;
 use crate::analyze::variables::GVar;
-use crate::generate::register_allocator::DataLocation;
-use crate::generate::register_allocator::RegId;
+use crate::generate::register_allocator::data_location::DataLocation;
+use crate::generate::register_allocator::data_location::RegId;
 use crate::preprocess::tokenkind::TokenKind as PreprocessTokenKind;
 use crate::tokenize::debug_infos::DebugInfo;
 use crate::tokenize::debug_infos::Position;
